--- src/main/java/webserver/ViewResolver.java
@@ -1,7 +1,0 @@
-package webserver;
-
-import java.io.IOException;
-
-public interface ViewResolver {
-    View resolve(final ServletRequest servletRequest) throws IOException;
-}
