+++ src/main/java/webserver/controller/IndexController.java
@@ -1,14 +1,9 @@
 package webserver.controller;
 
-import java.util.Collections;
-import java.util.LinkedHashMap;
-
 import com.google.common.collect.Maps;
-import webserver.MethodType;
-import webserver.ModelAndView;
-import webserver.RequestHeader;
-import webserver.ServletResponse;
-import webserver.StatusCode;
+import webserver.request.MethodType;
+import webserver.response.ModelAndView;
+import webserver.response.StatusCode;
 import webserver.controller.annotation.Controller;
 import webserver.controller.annotation.RequestMapping;
