+++ src/main/java/webserver/controller/annotation/RequestMapping.java
@@ -5,8 +5,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import webserver.MethodType;
-import webserver.RequestHeader;
+import webserver.request.MethodType;
 
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
