+++ src/main/java/webserver/request/MethodType.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import java.util.Arrays;
