+++ src/main/java/webserver/request/RequestBody.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
