+++ src/main/java/webserver/response/HttpStatusLine.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 public class HttpStatusLine {
     private String protocolVersion;
