+++ src/main/java/webserver/response/ServletResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 import java.util.Map;
