+++ src/test/java/webserver/RequestBodyTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import exception.InvalidRequestBodyException;
+import webserver.request.RequestBody;
 
 class RequestBodyTest {
     private static final String VALID_BODY = "userId=javajigi"
